cacesi/requirements.txt
Django==1.10.5
django-crispy-forms==1.6.1
django-filter==1.0.1
django-geoposition==0.3.0
djangorestframework==3.5.3
drf-nested-routers==0.11.1
Markdown==2.6.8
olefile==0.44
Pillow==4.0.0
